feat(theme): add dark/light mode toggle with persistence
- Create new ThemeContext for UI theme toggling (separate from DesignContext)
- Rename existing ThemeContext to DesignContext for design catalog feature
- Add theme toggle button to header using ThemeToggle component
- Implement localStorage persistence for theme preference
- Detect system preference (prefers-color-scheme) on first visit
- Add smooth CSS transition animation for theme changes
- Update Tailwind config to enable class-based dark mode
- Wire all components together using React context

Closes #68
