+++ src/components/ui/header.tsx
@@ -1,13 +1,10 @@
 import { Link } from 'react-router-dom'
 import { ThemeToggle } from './theme-toggle'
 import { LogoMark } from './LogoMark'
+import { useTheme } from '@/context/ThemeContext'
 
-interface HeaderProps {
-  theme: 'light' | 'dark'
-  onToggleTheme: () => void
-}
-
-export function Header({ theme, onToggleTheme }: HeaderProps) {
+export function Header() {
+  const { theme, toggleTheme } = useTheme()
   return (
     <header className="sticky top-0 z-50 w-full border-b bg-background/95 backdrop-blur supports-[backdrop-filter]:bg-background/60">
       <div className="container mx-auto flex h-16 items-center justify-between px-4 sm:px-6 lg:px-8">
